% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_normCentroidMz}
\alias{C_normCentroidMz}
\title{Rcpp function for normCentroidMz
SEXP}
\usage{
C_normCentroidMz(mz_vector, mz_upperLimit_vector)
}
\arguments{
\item{mz_vector}{A sorted numeric vecotor}

\item{mz_upperLimit_vector}{the mz_vector multiplied by ppmTol*(10^(-6)) and added to mz}
}
\description{
Rcpp function for normCentroidMz
SEXP
}
