% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mzDataTable.R
\name{.check_mzDataTable}
\alias{.check_mzDataTable}
\title{Check mzDt for data structure}
\usage{
.check_mzDataTable(mzDt)
}
\arguments{
\item{mzDt}{a data.table}
}
\value{

}
\description{
Check mzDt for data structure
}
