% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML2dataTable.R
\name{.formatHeader}
\alias{.formatHeader}
\title{Format header}
\usage{
.formatHeader(headerDF, removeEmptyCols = TRUE)
}
\arguments{
\item{headerDF}{the header returned by mzR::header()}

\item{removeEmptyCols}{Default is TRUE. Set FALSE if you want empty columns returned}
}
\value{
Returns a data.table
}
\description{
Format a header returned by mzR::header() by removing empty
columns, processing spectrumId column, and converting to data.table
}
\examples{

}
