% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBPI.R
\name{.getBPI_dt}
\alias{.getBPI_dt}
\title{Calculate total ion chromatogram from data.table}
\usage{
.getBPI_dt(mzDt)
}
\arguments{
\item{mzDt}{a data.table imported from mzML/mzXML file.}
}
\value{
returns a data.table.
}
\description{
Calculate total ion chromatogram from data.table
}
\details{
Internal function used by getTIC to calculate BPI from data.table.
}
