% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions_internals.R
\name{.getiRange}
\alias{.getiRange}
\title{Get index range from `seqNum`}
\usage{
.getiRange(mzObj, iStart, iStop)
}
\arguments{
\item{mzObj}{either a data.table or disk.frame.}

\item{iStart}{The integer start index to begin extracting ions.}

\item{iStop}{The integer stop index stop extracting ions.}
}
\value{
a vector of length two
}
\description{
Get index range from `seqNum`
}
