% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions_internals.R
\name{.getMzRange}
\alias{.getMzRange}
\title{Get mz start and stop as vector}
\usage{
.getMzRange(mz, mz_delta, ppmTol)
}
\arguments{
\item{mz}{the m/z value to extract.}

\item{mz_delta}{Optional. A numeric value +/- the mz value to extract.}

\item{ppmTol}{Optional. The mass accuracy of the instrument in ppm.}
}
\value{
a vector of length two specifying the start and stop m/z value.
}
\description{
Get mz start and stop as vector
}
