% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML2_OtherFormat_Internals.R
\name{.getNumFromSpectrumId}
\alias{.getNumFromSpectrumId}
\title{Convert spectrumId strings to numerics}
\usage{
.getNumFromSpectrumId(spectrumId)
}
\arguments{
\item{spectrumId}{the spectrumId character vector}
}
\value{
a data.frame with column names
}
\description{
Convert a character vector containing strings of the format
c("foobar=3 foo=45 bar=100") to a numeric data.frame
}
\examples{
\dontrun{
vec <- c("name1=10 name2=2 name3=333")
nums <- .getNumFromSpectrumId(spectrumId = vec)
nums
}

}
