% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSumSpectrum.R
\name{.getSumSpectrum_dskF}
\alias{.getSumSpectrum_dskF}
\title{Pre-process mzDskF for sum spectrum}
\usage{
.getSumSpectrum_dskF(mzDskF, iRange, tRange)
}
\arguments{
\item{mzDskF}{A disk.frame validated by .check_mzDataTable()}

\item{iRange}{seqNum range returned by .getiRange().}

\item{tRange}{retentionTime range returned by .gettRange().}
}
\value{
Return only mz and intensity as a data.table.
}
\description{
Pre-process mzDskF for sum spectrum
}
\details{
Extract mz and intensity from a disk.frame and optionally filter by
  time. Return only mz and intensity as a data.table.
}
