% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSumSpectrum.R
\name{.getSumSpectrum_dt}
\alias{.getSumSpectrum_dt}
\title{Pre-process mzDt for sum spectrum}
\usage{
.getSumSpectrum_dt(mzDt, iRange, tRange)
}
\arguments{
\item{mzDt}{a data.table validated by .check_mzDataTable()}

\item{iRange}{seqNum range returned by .getiRange().}

\item{tRange}{retentionTime range returned by .gettRange().}
}
\value{
Return only mz and intensity as a data.table.
}
\description{
Pre-process mzDt for sum spectrum
}
\details{
Extract mz and intensity from a data.table and optionally filter by
  time. Return only mz and intensity as a data.table.
}
