% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTIC_dt.R
\name{.getTIC_dskF}
\alias{.getTIC_dskF}
\title{Calculate total ion chromatogram from disk.frame}
\usage{
.getTIC_dskF(mzDskF)
}
\arguments{
\item{mzDskF}{a disk.frame imported from mzML/mzXML file.}
}
\value{
returns a data.table.
}
\description{
Calculate total ion chromatogram from disk.frame
}
\details{
Internal function used by getTIC to calculate TIC from disk.frame.
}
