% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXIC.R
\name{.gettRange}
\alias{.gettRange}
\title{Get time range from `retentionTime`}
\usage{
.gettRange(mzObj, tStart, tStop)
}
\arguments{
\item{mzObj}{Either a data.table or disk.frame.}

\item{tStart}{The numeric time value to begin extracting ions.}

\item{tStop}{The numeric time value to stop extracting ions.}
}
\value{
a vector of length two
}
\description{
Get time range from `retentionTime`
}
