% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXIC.R
\name{.getXIC_dskF}
\alias{.getXIC_dskF}
\title{Get XIC from disk.frame}
\usage{
.getXIC_dskF(mzDskF, mzRange, iRange, tRange)
}
\arguments{
\item{mzDskF}{A disk.frame validated by .check_mzDataTable()}

\item{mzRange}{m/z range returned by .getMzRange()}

\item{iRange}{seqNum range returned by .getiRange()}

\item{tRange}{retentionTime range returned by .gettRange()}
}
\value{
Returns a data.table
}
\description{
Get XIC from disk.frame
}
\details{
non-exported function for use in getXIC.
}
