% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXIC.R
\name{.getXIC_dt}
\alias{.getXIC_dt}
\title{Get XIC from data.table}
\usage{
.getXIC_dt(mzDt, mzRange, iRange, tRange)
}
\arguments{
\item{mzDt}{a data.table validated by .check_mzDataTable()}

\item{mzRange}{m/z range returned by .getMzRange()}

\item{iRange}{seqNum range returned by .getiRange()}

\item{tRange}{retentionTime range returned by .gettRange()}
}
\value{
Returns a data.table
}
\description{
Get XIC from data.table
}
\details{
non-exported function for use in getXIC.
}
