% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSimilarity.R
\name{.grid_matrix}
\alias{.grid_matrix}
\title{Grid spectra to equal lengths}
\usage{
.grid_matrix(mzDt, seqNums)
}
\arguments{
\item{mzDt}{A data.table minimally containing "intensity" and "mzGrid_index"}

\item{seqNums}{Original spectra index values "seqNum" imported from mzML/mzXML files.}
}
\value{
a matrix
}
\description{
After global mz gridding of n-spectra with
  indexMasterSpectrum(), fill matrix containing n-columns to generate
  comparable spectra.
}
