% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions_internals.R
\name{.inGlobalEnv}
\alias{.inGlobalEnv}
\title{Check if all objects are in .GlobalEnv}
\usage{
.inGlobalEnv(varNames)
}
\arguments{
\item{varNames}{a character vector of any length providing variable names}
}
\value{
logical vector of length one
}
\description{
Check if all objects are in .GlobalEnv
}
