% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML2csv.R
\name{.mzML2csvChunk}
\alias{.mzML2csvChunk}
\title{Write data.table of mzML/mzXML data to a .csv}
\usage{
.mzML2csvChunk(path, csvPath, scans)
}
\arguments{
\item{path}{path or mzR pointer to the mzML file}

\item{csvPath}{path to a .csv file}

\item{scans}{Provide a numeric vector to import select scans and write to a
.csv file. If not provided, the whole file is converted. Likely provided by
.scanChunker()}
}
\description{
Internal function. Used by mzML2csv in an mapply function.
  Import data using mzML2dataTable() and write results to .csv, appending
  results after the first import.
}
