% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML2diskFrame.R
\name{.mzML2diskFrameChunk}
\alias{.mzML2diskFrameChunk}
\title{Write data.table of mzML/mzXML data to a disk.frame}
\usage{
.mzML2diskFrameChunk(path, diskFrame, scans)
}
\arguments{
\item{path}{path or mzR pointer to the mzML file}

\item{diskFrame}{disk.frame pointer object created by disk.frame()}

\item{scans}{a numeric vector to import select scans and write to a
data.frame - likely generated by .scanChunker()}
}
\description{
Internal function. Import data using mzML2dataTable() and write
  results to a disk.frame as a chunk.
}
