% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeSpectrum_dt.R
\name{.normalizeEachSpectrum_dt}
\alias{.normalizeEachSpectrum_dt}
\title{Normalize a mzDt by seqNum}
\usage{
.normalizeEachSpectrum_dt(mzDt, method)
}
\arguments{
\item{mzDt}{the mzDt to be normalized in place}

\item{method}{"maxPeak", "sum", or "sqrt"}
}
\value{
returns mzDt with extra column `intensity_norm`
}
\description{
Normalize intensity of each individual spectrum in the mzDt.
}
\details{
"maxPeak": divide intensity of each spectrum by the max intensity.
  "sqrt": divide intensity of each spectrum using sqrt intensity transform.
  "sum" : divide intensity of each spectrum by sum intensity of the spectrum.
}
