% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeSpectrum_dt.R
\name{.normalizeSpectrum_dt}
\alias{.normalizeSpectrum_dt}
\title{Normalize a mzDt}
\usage{
.normalizeSpectrum_dt(mzDt)
}
\arguments{
\item{mzDt}{the mzDt to be normalized in place}
}
\value{
returns mzDt with extra column `intensity_norm`
}
\description{
Normalize a mzDt
}
