% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexMasterSpectrum.R
\name{.normProfileMz}
\alias{.normProfileMz}
\title{Calculate global m/z grid for profile mode data}
\usage{
.normProfileMz(mz)
}
\arguments{
\item{mz}{a vector of m/z values to be clustered}
}
\value{
a data.table
}
\description{
Calculate global m/z grid for profile mode data
}
