% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML2_OtherFormat_Internals.R
\name{.scanChunker}
\alias{.scanChunker}
\title{Scan Chunker}
\usage{
.scanChunker(scans, mzRfilePointer, chunkSize)
}
\arguments{
\item{scans}{A numeric vector of all of the scan indices to be imported}

\item{mzRfilePointer}{A mzR pointer object linking to the file to be
converted}

\item{chunkSize}{The number of scans to be processed at a time - actual
import size will be aproximately the size of chunkSize}
}
\value{
a list object containing the group indices and order of those indicies. See .splitIndex.
}
\description{
Break sets of ms scans into indexed chunks for multi-import methods
}
\examples{
\dontrun{
chunks <- .scanChunker(scans = c(1:10), chunkSize = 2)
}

}
