% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSimilarity.R
\name{.specSim}
\alias{.specSim}
\title{Calculate spectral similarity}
\usage{
.specSim(mzDt, similarity_method)
}
\arguments{
\item{mzDt}{the data.table containing two or more spectra to be compared.}

\item{similarity_method}{choose one: "dotProd", "specContrast", "specCor",
"brayCurtis", "euclidean"}
}
\value{
returns a 3-column matrix containing the similarity score and
  sequence numbers
}
\description{
Calculate the spectral similarity using one of the available
  methods. The mzDt containing n-unique seqNums is converted into a matrix
  with n-columns. All unique combinations are calculated and returned. The
  output can be further processed to a similarity matrix using
  similarity2SquareMatrix(). mzDt should contain self-normalized spectra
  (see: .normalizeEachSpectrum_dt()). Each spectrum should contain mzGrid and
  mzGrid_index corresponding to the global mz grid (see:
  indexMasterSpectrum()).
}
