% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML2_OtherFormat_Internals.R
\name{.splitDFbyVectorList}
\alias{.splitDFbyVectorList}
\title{select data.frame by row}
\usage{
.splitDFbyVectorList(df, vectorList)
}
\arguments{
\item{df}{a data frame to be split by row index}

\item{vectorList}{rows to return}
}
\value{
retuns a data.frame/data.table subset
}
\description{
Internal function to be used by mapply for splitting a data.frame/data.table
by row using a vector of index values
}
\examples{
\dontrun{
dat <- data.frame(col1 = c(1:10), col2 = c(21:30))
vecList <- list(c(1:3), c(6:8), c(1:5))
split.dat <- mapply(FUN = .splitDFbyVectorList,
                   vectorList = vecList,
                   MoreArgs = list(dat),
                   SIMPLIFY = FALSE)
}

}
