% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitIndex.R
\name{.splitIndex}
\alias{.splitIndex}
\title{Split a vector into groups}
\usage{
.splitIndex(index, nGroups, randomize = FALSE)
}
\arguments{
\item{index}{the vector of index values to be split into groups}

\item{nGroups}{number of groups to be generated}

\item{randomize}{should the groups contain a random or ordered sampling from
the index vector}
}
\value{
returns a list containing `groups` and `groupIndex` both of which are
  lists of length nGroups
}
\description{
Split a vector into groups
}
\examples{
\dontrun{
index <- c(c(1:10), c(30:40))
groupedIndex <- .splitIndex(nGroups = 3, index = index, randomize = TRUE)

groups <- unlist(groupedIndex$groups)
groups_i <- unlist(groupedIndex$groupIndex)

all(index \%in\% groups)

#Reorder the radomized groups by groups_i
groups_reorder <- vector(mode = "numeric", length = length(groups))
groups_reorder[groups_i] <- groups
groups_reorder}

}
