% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBPI_dt.R
\name{getBPI_dt}
\alias{getBPI_dt}
\title{Get Base Peak Intensity}
\usage{
getBPI_dt(mzDt, normalize = FALSE)
}
\arguments{
\item{mzDt}{a data.table imported by mzML2dataTable().}

\item{normalize}{}
}
\value{

}
\description{
Get Base Peak Intensity
}
\examples{

}
