% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSumSpectrum.R
\name{getSumSpectrum}
\alias{getSumSpectrum}
\title{Get Sum Mass Spectrum}
\usage{
getSumSpectrum(
  mzObj,
  ppmTol = NULL,
  iStart = NULL,
  iStop = NULL,
  tStart = NULL,
  tStop = NULL,
  normalize = FALSE,
  isCentroid = TRUE
)
}
\arguments{
\item{mzObj}{a data.table or disk.frame imported from mzML/mzXML file.}

\item{ppmTol}{The mass accuracy of the instrument in ppm. Default =
NULL}

\item{iStart}{Optional. The integer start index to begin extracting ions.
Default = NULL.}

\item{iStop}{Optional. The integer stop index stop extracting ions. Default =
NULL.}

\item{tStart}{Optional. The numeric time value to begin extracting ions.
Default = NULL.}

\item{tStop}{Optional. The numeric time value to stop extracting ions.
Default = NULL.}

\item{normalize}{Logical. Default = `FALSE`. Should the data be normalized?}

\item{isCentroid}{Logical. The default is `TRUE`. Are the spectra centroids?}
}
\value{
Returns a data.table
}
\description{
Sum multiple scans and optionally normalize the spectrum. Scans
  are gridded by the grouping algorithm `indexMasterSpectrum()`.
}
\examples{

}
