% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSumSpectrum_dt.R
\name{getSumSpectrum_dt}
\alias{getSumSpectrum_dt}
\title{Get Sum Mass Spectrum}
\usage{
getSumSpectrum_dt(
  mzDt,
  ppmTol = NULL,
  iStart = NULL,
  iStop = NULL,
  tStart = NULL,
  tStop = NULL,
  normalize = FALSE,
  isCentroid = TRUE
)
}
\arguments{
\item{mzDt}{a data.table imported by `mzML2dataTable()`.}

\item{iStart}{Optional. The integer start index to begin extracting ions.
Default = NULL.}

\item{iStop}{Optional. The integer stop index stop extracting ions. Default =
NULL.}

\item{tStart}{Optional. The numeric time value to begin extracting ions.
Default = NULL.}

\item{tStop}{Optional. The numeric time value to stop extracting ions.
Default = NULL.}

\item{normalize}{Logical. Default = `FALSE`. Should the data be normalized?}

\item{isCentroid}{Logical. The default is `TRUE`. Are the spectra centroids?}

\item{ppmTolOptional.}{The mass accuracy of the instrument in ppm. Default =
NULL}
}
\value{
Returns a data.table
}
\description{
Sum multiple scans and optionally normalize the spectrum. Scans
  are gridded by the grouping algorithm `indexMasterSpectrum()`.
}
\examples{

}
