% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTIC.R
\name{getTIC}
\alias{getTIC}
\title{Calculate total ion chromatogram}
\usage{
getTIC(mzObj, normalize = FALSE, method = "sqrt")
}
\arguments{
\item{mzObj}{a data.table or disk.frame imported from mzML/mzXML file.}

\item{normalize}{Logical. Default = FALSE. Should the data be normalized?}

\item{method}{"maxPeak", "sum", or "sqrt"}
}
\value{

}
\description{
Calculate total ion chromatogram
}
\examples{

}
