% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTIC_dt.R
\name{getTIC_dt}
\alias{getTIC_dt}
\title{Calculate total ion chromatogram}
\usage{
getTIC_dt(mzDt, normalize = FALSE)
}
\arguments{
\item{mzDt}{a data.table imported by mzML2dataTable().}

\item{normalize}{Logical. Default = FALSE. Should the data be normalized?}
}
\value{
returns a data.table possibly containing normalized intensity.
}
\description{
Calculate total ion chromatogram
}
\examples{

}
