% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXIC.R
\name{getXIC}
\alias{getXIC}
\title{Get XIC from data.table or disk.frame}
\usage{
getXIC(
  mzObj,
  mz,
  mz_delta = NULL,
  ppmTol = NULL,
  iStart = NULL,
  iStop = NULL,
  tStart = NULL,
  tStop = NULL
)
}
\arguments{
\item{mzObj}{a data.table or disk.frame imported from mzML/mzXML file.}

\item{mz}{the m/z value to extract.}

\item{mz_delta}{Optional. A numeric value +/- the mz value to extract.
Default = NULL.}

\item{ppmTol}{Optional. The mass accuracy of the instrument in ppm.
Default = NULL}

\item{iStart}{Optional. The integer start index to begin extracting ions.
Default = NULL.}

\item{iStop}{Optional. The integer stop index stop extracting ions. Default =
NULL.}

\item{tStart}{Optional. The numeric time value to begin extracting ions.
Default = NULL.}

\item{tStop}{Optional. The numeric time value to stop extracting ions.
Default = NULL.}
}
\value{
Returns a subset of the original mzObj as a data.table
}
\description{
Extract ions by mz using a variety of options for subsetting.
  See details.
}
\details{
mzObj and mz must be provided. Default behavior is to extract ions
  exactly matching mz value accross all time.

  Specifying mz_delta or ppmTol will result in mz filtering. mz_delta
  will extract ions with values in the range where mz >= mz-mz_delta and mz
  <= mz+mz_delta. Specifying ppmTol workes similarly, but calculates
  the upper and lower bounds from ppm mass error.

  Specifying iStart/iStop or tStart/tStop will result in time filtering.
}
\examples{

}
