% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXIC_dt.R
\name{getXIC_dt}
\alias{getXIC_dt}
\title{Get XIC from mzDataTable}
\usage{
getXIC_dt(
  mzDt,
  mz,
  mz_delta = NULL,
  massAccuracy = NULL,
  iStart = NULL,
  iStop = NULL,
  tStart = NULL,
  tStop = NULL
)
}
\arguments{
\item{mzDt}{a data.table imported by mzML2dataTable().}

\item{mz}{the m/z value to extract.}

\item{mz_delta}{Optional. A numeric value +/- the mz value to extract.
Default = NULL.}

\item{massAccuracy}{Optional. The mass accuracy of the instrument in ppm.
Default = NULL}

\item{iStart}{Optional. The integer start index to begin extracting ions.
Default = NULL.}

\item{iStop}{Optional. The integer stop index stop extracting ions. Default =
NULL.}

\item{tStart}{Optional. The numeric time value to begin extracting ions.
Default = NULL.}

\item{tStop}{Optional. The numeric time value to stop extracting ions.
Default = NULL.}
}
\value{
Returns a subset of the original data.table
}
\description{
Extract ions by mz using a variety of options for subsetting.
  See details.
}
\details{
mzDt and mz must be provided. Default behavior is to extract ions
  exactly matching mz value provided accross all time.

  Specifying mz_delta or massAccuracy will result in mz filtering. mz_delta
  will extract ions with values in the range where mz >= mz-mz_delta and mz
  <= mz+mz_delta. Specifying massAccuracy workes similarly, but calculates
  the upper and lower bounds from ppm error.

  Specifying iStart/iStop or tStart/tStop will result in time filtering.
}
\examples{

}
