% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexMasterSpectrum.R
\name{indexMasterSpectrum}
\alias{indexMasterSpectrum}
\title{Index the m/z axis}
\usage{
indexMasterSpectrum(dt, ppmTol, isCentroid = TRUE)
}
\arguments{
\item{dt}{a data.table containing a column named mz}

\item{ppmTol}{the grouping tolerance in ppm}

\item{isCentroid}{Logical. The default is `TRUE`. Are the spectra centroids?}
}
\value{
Returns the original data.table also containing the the normalized
  m/z values and their index
}
\description{
Index the m/z axis for a set of mass spectra. Function outputs
  normalized m/z values and integer index values.
}
\examples{
mz <- c(100, 100.01, 100.02, 100.021, 100.2, 100.21, 100.22, 100.22, 100.31)

}
