% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML2csv.R
\name{mzML2csv}
\alias{mzML2csv}
\title{Convert .mzML file to CSV using mzR and data.table}
\usage{
mzML2csv(path, outputPath, scans = NULL, chunkSize = 100)
}
\arguments{
\item{path}{path to the mzML file}

\item{outputPath}{path and fileName.csv specifying write location of .csv
file}

\item{scans}{Optional parameter. Provide a numeric vector to import select
scans and write to a .csv file. If not provided, the whole file is
converted. Default is NULL.}

\item{chunkSize}{number of scans to be extracted and written to a .csv file
at a time. Useful for breaking up large data files and converting to .csv.
Default is 100. If NULL, imports all scans in single operation and you will
probably run out of memory.}
}
\value{
the results from writing the .csv file
}
\description{
Convert .mzML file to CSV using mzR and data.table
}
\examples{

}
