% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML2dataTable.R
\name{mzML2dataTable}
\alias{mzML2dataTable}
\title{Import a mzML file and format the data into a data.table}
\usage{
mzML2dataTable(path, scans = NULL, header = NULL)
}
\arguments{
\item{path}{Either a system path to the .mzML file or mzR pointer object}

\item{scans}{A numeric vecotr specifying which scans to return. Optional
argument. If ommited, all peaks are returned. Default is `NULL`}

\item{header}{Option to provide a header data.frame/data.table if it has
already been imported. Expects header spans range of `scans`. Default is
NULL.}
}
\value{
a data.table
}
\description{
Convert mzML/mzXML data files into data.tables. This function is
  used by other converters in this package and provides flexible inputs - see
  parameters for details.
}
\examples{
\dontrun{
#Typical use for user - import mzML/mzXML data as data.table
dt <- mzML2dataTable(path = "mzML_path.mzML", scans = c(1:100))

#Typical use for other functions in this package
#-import mzML/mzXML data as data.table from mzR pointer
mzML <- mzR::openMSfile(filename = "mzML_path.mzML")
dt <- mzML2dataTable(path = mzML, scans = c(1:100))

#-import mzML/mzXML scan data with previously imported header
header <- mzR::header(object = mzML, scans = c(1:100))
dt <- mzML2dataTable(path = mzML,
                     scans = c(1:100),
                     header = header)
}

}
