% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML2dataTable_par.R
\name{mzML2dataTable_par}
\alias{mzML2dataTable_par}
\title{parallalized mzML import - does not work :(}
\usage{
mzML2dataTable_par(path, scans = NULL, nthreads = NULL)
}
\arguments{
\item{path}{a system path to the .mzML file}

\item{scans}{A numeric specifying which scans to return. Optional argument.
If ommited, the requested data for all peaks is returned.}

\item{nthreads}{the number of threads to use for parallel read}
}
\value{
a data.table
}
\description{
parallalized mzML import - does not work :(
}
\examples{

}
