% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSimilarity.R
\name{similarity2SquareMatrix}
\alias{similarity2SquareMatrix}
\title{Generate square similarity matrix}
\usage{
similarity2SquareMatrix(simMat)
}
\arguments{
\item{simMat}{the matrix returned by computeSimilarity() or .specSim() containing columns "sim", "seqNum1", "seqNum2"}
}
\value{
Returns a square matrix
}
\description{
Generate square similarity matrix
}
\examples{
#See computeSimilarity
}
