% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mzDataTable.R
\name{.check_mzDataTable}
\alias{.check_mzDataTable}
\title{Check mzDt for data structure}
\usage{
.check_mzDataTable(mzDt)
}
\arguments{
\item{mzDt}{a data.table or a disk.frame}
}
\value{
return `TRUE` for data.table and `FALSE` for disk.frame.
}
\description{
Check mzDt for data structure
}
\details{
Function is intended to check for expected column types in mzObj
  passed to various processing functions. It will return `TRUE` for
  data.table and `FALSE` for disk.frame and throw an error if either are not
  properly formatted.
}
