% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mzDataTable.R
\name{.check_mzDataTable}
\alias{.check_mzDataTable}
\title{Check mzObj for data structure}
\usage{
.check_mzDataTable(mzObj)
}
\arguments{
\item{mzObj}{a data.table or disk.frame imported from mzML/mzXML file.}
}
\value{
return `TRUE` for data.table and `FALSE` for disk.frame.
}
\description{
Check mzObj for data structure
}
\details{
Function is intended to check for expected column types in mzObj
  passed to various processing functions. It will return `TRUE` for
  data.table and `FALSE` for disk.frame and throw an error if either are not
  properly formatted.
}
