% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeSpectrum_dt.R
\name{.normalizeSpectrum_dt}
\alias{.normalizeSpectrum_dt}
\title{Normalize a mzDt}
\usage{
.normalizeSpectrum_dt(mzDt, method)
}
\arguments{
\item{mzDt}{the mzDt to be normalized in place}

\item{method}{"maxPeak", "sum", or "sqrt"}
}
\value{
returns mzDt with extra column `intensity_norm`
}
\description{
Normalize intensity of mzDt by one of the available methods.
  Normalization is global rather than by scan.
}
\details{
"maxPeak": divide intensity by the max intensity. "sqrt": divide
  intensity using sqrt intensity transform. "sum" : divide intensity by sum
  intensity.
}
