% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML2dataTable.R
\name{mzML2dataTable}
\alias{mzML2dataTable}
\title{Import a mzML file and format the data into a data.table}
\usage{
mzML2dataTable(path, scans = NULL, header = NULL)
}
\arguments{
\item{path}{Either a system path to the .mzML file or mzR pointer object}

\item{scans}{A numeric vecotr specifying which scans to return. Optional
argument. If ommited, all peaks are returned. Default is `NULL`}

\item{header}{Option to provide a header data.frame/data.table if it has
already been imported. Expects header spans range of `scans`. Default is
NULL.}
}
\value{
a data.table
}
\description{

}
\examples{

}
