% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML2diskFrame.R
\name{mzML2diskFrame}
\alias{mzML2diskFrame}
\title{Convert mz data to disk.frame}
\usage{
mzML2diskFrame(path, diskFramePath, scans = NULL, chunkSize = NULL)
}
\arguments{
\item{path}{path to the mzML file}

\item{diskFramePath}{path and fileName.csv specifying write location of .csv
file}

\item{scans}{Optional parameter. Provide a numeric vector to import select
scans and write to a .csv file. If not provided, the whole file is
converted. Default is NULL.}

\item{chunkSize}{number of scans to be extracted and written to a .csv file
at a time. Useful for breaking up large data files and converting to .csv.
Default is 100. If NULL, imports all scans in single operation and you will
probably run out of memory.}
}
\value{
Returns a disk.frame reference object.
}
\description{
Convert .mzML file to disk.frame using mzR, data.table, and disk.frame
}
\examples{

}
